Get generated projects closer to shipping SDKs
