+++ AutoRest.CSharp.V3.sln
@@ -14,8 +14,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AutoRest.CoreShared", "src\AutoRest.CoreShared\AutoRest.CoreShared.csproj", "{9DB44A43-CFD6-4868-844F-70A25834AE3D}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -34,10 +32,6 @@ Global
 		{C0EC4E09-EB17-4C6B-AB31-F14E5C834D7D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C0EC4E09-EB17-4C6B-AB31-F14E5C834D7D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C0EC4E09-EB17-4C6B-AB31-F14E5C834D7D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9DB44A43-CFD6-4868-844F-70A25834AE3D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9DB44A43-CFD6-4868-844F-70A25834AE3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9DB44A43-CFD6-4868-844F-70A25834AE3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9DB44A43-CFD6-4868-844F-70A25834AE3D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
