+++ eng/CodeChecks.ps1
@@ -46,6 +46,11 @@ function Invoke-Block([scriptblock]$cmd) {
 }
 
 try {
+    Write-Host "Downloading files"
+    Invoke-Block {
+        & $PSScriptRoot\DownloadSharedSource.ps1 @script:PSBoundParameters
+    }
+
     Write-Host "Generate test clients"
     Invoke-Block {
         & $PSScriptRoot\Generate.ps1 @script:PSBoundParameters
