--- src/AutoRest.CSharp.V3/Azure.Core.Shared/DiagnosticScope.cs
@@ -1,157 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-#nullable enable
-
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Reflection;
-
-namespace Azure.Core.Pipeline
-{
-    public readonly struct DiagnosticScope : IDisposable
-    {
-        private readonly DiagnosticActivity? _activity;
-
-        private readonly string _name;
-
-        private readonly DiagnosticListener _source;
-
-        internal DiagnosticScope(string name, DiagnosticListener source)
-        {
-            _name = name;
-            _source = source;
-            _activity = _source.IsEnabled(_name) ? new DiagnosticActivity(_name) : null;
-            _activity?.SetW3CFormat();
-        }
-
-        public bool IsEnabled => _activity != null;
-
-        public void AddAttribute(string name, string value)
-        {
-            _activity?.AddTag(name, value);
-        }
-
-        public void AddAttribute<T>(string name, T value)
-        {
-            if (_activity != null && value != null)
-            {
-                AddAttribute(name, value.ToString()!);
-            }
-        }
-
-        public void AddAttribute<T>(string name, T value, Func<T, string> format)
-        {
-            if (_activity != null)
-            {
-                AddAttribute(name, format(value));
-            }
-        }
-
-        public void AddLink(string id)
-        {
-            if (_activity != null)
-            {
-                var linkedActivity = new Activity("LinkedActivity");
-                linkedActivity.SetW3CFormat();
-                linkedActivity.SetParentId(id);
-
-                _activity.AddLink(linkedActivity);
-            }
-        }
-
-        public void Start()
-        {
-            if (_activity != null)
-            {
-                _source.StartActivity(_activity, _activity);
-            }
-        }
-
-        public void Dispose()
-        {
-            if (_activity == null)
-            {
-                return;
-            }
-
-            if (_source != null)
-            {
-                _source.StopActivity(_activity, null);
-            }
-            else
-            {
-                _activity?.Stop();
-            }
-        }
-
-        public void Failed(Exception e)
-        {
-            if (_activity == null)
-            {
-                return;
-            }
-
-            _source?.Write(_activity.OperationName + ".Exception", e);
-
-        }
-
-        private class DiagnosticActivity : Activity
-        {
-            private List<Activity>? _links;
-
-            public IEnumerable<Activity> Links => (IEnumerable<Activity>?)_links ?? Array.Empty<Activity>();
-
-            public DiagnosticActivity(string operationName) : base(operationName)
-            {
-            }
-
-            public void AddLink(Activity activity)
-            {
-                _links ??= new List<Activity>();
-                _links.Add(activity);
-            }
-        }
-    }
-
-    /// <summary>
-    /// HACK HACK HACK. Some runtime environments like Azure.Functions downgrade System.Diagnostic.DiagnosticSource package version causing method not found exceptions in customer apps
-    /// This type is a temporary workaround to avoid the issue.
-    /// </summary>
-    internal static class ActivityExtensions
-    {
-        private static readonly MethodInfo? s_setIdFormatMethod = typeof(Activity).GetMethod("SetIdFormat");
-        private static readonly MethodInfo? s_getIdFormatMethod = typeof(Activity).GetProperty("IdFormat")?.GetMethod;
-        private static readonly MethodInfo? s_getTraceStateStringMethod = typeof(Activity).GetProperty("TraceStateString")?.GetMethod;
-
-        public static bool SetW3CFormat(this Activity activity)
-        {
-            if (s_setIdFormatMethod == null) return false;
-
-            s_setIdFormatMethod.Invoke(activity, new object[]{ 2 /* ActivityIdFormat.W3C */});
-
-            return true;
-        }
-
-        public static bool IsW3CFormat(this Activity activity)
-        {
-            if (s_getIdFormatMethod == null) return false;
-
-            object? result = s_getIdFormatMethod.Invoke(activity, Array.Empty<object>());
-
-            return result is int resultInt && resultInt == 2 /* ActivityIdFormat.W3C */;
-        }
-
-        public static bool TryGetTraceState(this Activity activity, out string? traceState)
-        {
-            traceState = null;
-
-            if (s_getTraceStateStringMethod == null) return false;
-
-            traceState = s_getTraceStateStringMethod.Invoke(activity, Array.Empty<object>()) as string;
-
-            return true;
-        }
-    }
-}
