+++ src/AutoRest.CSharp.V3/CodeGen/CsProjWriter.cs
@@ -10,37 +10,14 @@ namespace AutoRest.CSharp.V3.CodeGen
             Line($@"<Project Sdk=""Microsoft.NET.Sdk"">
 
   <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <AssemblyName>{configuration.Namespace}</AssemblyName>
-    <RootNamespace>{configuration.Namespace}</RootNamespace>
-    <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-    <OutputPath>bin</OutputPath>
-    <PublishDir>$(OutputPath)</PublishDir>
-    <NoWarn>SA1649,CS8618</NoWarn>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <WarningsAsErrors />
     <Nullable>annotations</Nullable>
-    <RunAnalyzersDuringBuild>true</RunAnalyzersDuringBuild>
-    <RunAnalyzersDuringLiveAnalysis>false</RunAnalyzersDuringLiveAnalysis>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=""'$(Configuration)|$(Platform)'=='Debug|AnyCPU'"">
-    <DelaySign>false</DelaySign>
-    <DefineConstants>TRACE;DEBUG;NETSTANDARD</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=""'$(Configuration)|$(Platform)'=='Release|AnyCPU'"">
-    <SignAssembly>true</SignAssembly>
-    <DelaySign>true</DelaySign>
-    <AssemblyOriginatorKeyFile>MSSharedLibKey.snk</AssemblyOriginatorKeyFile>
-    <DefineConstants>TRACE;RELEASE;NETSTANDARD;SIGN</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include=""Azure.Core"" Version=""1.0.0"" />
-    <PackageReference Include=""Azure.Identity"" Version=""1.0.0"" />
-    <PackageReference Include=""NUnit"" Version=""3.12.0"" />
+    <PackageReference Include=""System.Text.Json"" Version=""4.6.0"" />
   </ItemGroup>
 
 </Project>
