+++ src/assets/Azure.Core.Shared/ArrayBufferWriter.cs
@@ -10,7 +10,7 @@ namespace Azure.Core
     /// <summary>
     /// Represents a heap-based, array-backed output sink into which <typeparam name="T"/> data can be written.
     /// </summary>
-    public sealed class ArrayBufferWriter<T> : IBufferWriter<T>
+    internal sealed class ArrayBufferWriter<T> : IBufferWriter<T>
     {
         private T[] _buffer;
         private const int DefaultInitialBufferSize = 256;
